[feat] Introduce Screen class for displaying roulette results
Added a Screen class to the basic roulette game implementation. The main
script now creates an instance of the Screen class to display the ball pocket
number to the user after spinning the wheel. This improves the user interface
by separating the logic of the game from the display functionality.